# code/service/hrms/hrms-openapi.yaml
openapi: 3.0.3
info:
  title: DIGIT HRMS – Employee Service API
  version: 3.0.0
  description: |
    REST APIs to manage employees in the DIGIT HRMS system.
    This module provides CRUD and lifecycle operations for employees,
    including jurisdictions, assignments, service history, education, and document records.
servers:
  - url: https://api.example.com/
    description: Production
  - url: https://staging.example.com/
    description: Staging
security:
  - BearerAuth: []
...
paths:
...
  /employees/v3/jurisdictions:
    post:
      tags: [Jurisdictions]
      summary: Create a jurisdiction
      operationId: createJurisdiction
      description: |
        Creates a new jurisdiction for a tenant.
        `tenantId` in payload is ignored and derived from the header.
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/ClientId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jurisdiction]
              properties:
                jurisdiction:
                  $ref: '#/components/schemas/Jurisdiction'
      responses:
        '201':
          description: Jurisdiction created
          headers:
            X-Response-Time: { $ref: '#/components/headers/X-Response-Time' }
            X-Response-Timestamp: { $ref: '#/components/headers/X-Response-Timestamp' }
            X-Request-ID: { $ref: '#/components/headers/X-Request-ID' }
            X-Correlation-ID: { $ref: '#/components/headers/X-Correlation-ID' }
            X-Tenant-ID: { $ref: '#/components/headers/X-Tenant-ID' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Jurisdiction'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Error' }
        '409':
          description: Duplicate jurisdiction record
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    get:
      tags: [Jurisdictions]
      summary: Search jurisdictions
      operationId: searchJurisdictions
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/ClientId'
        - in: query
          name: ids
          schema:
            type: array
            items: { type: string, format: uuid }
        - in: query
          name: employeeId
          schema: { type: string, format: uuid }
          description: Filter by employee UUID

        - in: query
          name: employeeIds
          schema:
            type: array
            items: { type: string, format: uuid }
          style: form
          explode: true
          description: Filter multiple employees' jurisdictions

        - in: query
          name: isActive
          schema: { type: boolean }
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Jurisdictions found
          headers:
            X-Response-Time: { $ref: '#/components/headers/X-Response-Time' }
            X-Response-Timestamp: { $ref: '#/components/headers/X-Response-Timestamp' }
            X-Request-ID: { $ref: '#/components/headers/X-Request-ID' }
            X-Correlation-ID: { $ref: '#/components/headers/X-Correlation-ID' }
            X-Tenant-ID: { $ref: '#/components/headers/X-Tenant-ID' }
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Jurisdiction'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Error' }
  /employees/v3/jurisdictions/{uuid}:
    get:
      tags: [Jurisdictions]
      summary: Get jurisdiction by UUID
      operationId: getJurisdictionByUUID
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/ClientId'
        - name: uuid
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Jurisdiction found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Jurisdiction' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    put:
      tags: [Jurisdictions]
      summary: Replace jurisdiction
      operationId: replaceJurisdiction
      parameters:
        - $ref: '#/components/parameters/TenantIdHeader'
        - $ref: '#/components/parameters/ClientId'
        - name: uuid
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [jurisdiction]
              properties:
                jurisdiction:
                  $ref: '#/components/schemas/Jurisdiction'
      responses:
        '200':
          description: Updated successfully
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Jurisdiction' }
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Error' }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
components:
  securitySchemes:
    BearerAuth:
      $ref: 'https://raw.githubusercontent.com/egovernments/DIGIT-3.0/master/common-3.0.0.yaml#/components/securitySchemes/BearerAuth'
  parameters:
...
    TenantIdHeader:
      $ref: 'https://raw.githubusercontent.com/digitnxt/digit3/master/docs/services/common/Common-3.0.0.yaml#/components/parameters/TenantIdHeader'
...
    ClientId:
      $ref: 'https://raw.githubusercontent.com/digitnxt/digit3/master/docs/services/common/Common-3.0.0.yaml#/components/parameters/ClientId'
...
  headers:
    X-Response-Time:
      $ref: 'https://raw.githubusercontent.com/digitnxt/digit3/master/docs/services/common/Common-3.0.0.yaml#/components/headers/X-Response-Time'
    X-Response-Timestamp:
      $ref: 'https://raw.githubusercontent.com/digitnxt/digit3/master/docs/services/common/Common-3.0.0.yaml#/components/headers/X-Response-Timestamp'
    X-Request-ID:
      $ref: 'https://raw.githubusercontent.com/digitnxt/digit3/master/docs/services/common/Common-3.0.0.yaml#/components/headers/X-Request-ID'
    X-Correlation-ID:
      $ref: 'https://raw.githubusercontent.com/digitnxt/digit3/master/docs/services/common/Common-3.0.0.yaml#/components/headers/X-Correlation-ID'
    X-Tenant-ID:
      $ref: 'https://raw.githubusercontent.com/digitnxt/digit3/master/docs/services/common/Common-3.0.0.yaml#/components/headers/X-Tenant-ID'
  schemas:
    AuditDetails:
      $ref: 'https://raw.githubusercontent.com/digitnxt/digit3/refs/heads/master/docs/services/common/Common-3.0.0.yaml#/components/schemas/AuditDetail'
    Error:
      $ref: 'https://raw.githubusercontent.com/digitnxt/digit3/master/docs/services/common/Common-3.0.0.yaml#/components/schemas/Error'
...
    Jurisdiction:
      type: object
      description: Represents an organisational jurisdiction within the HRMS system.
      required:
        - employeeId
        - boundaryRelation
      properties:
        id:
          type: string
          format: uuid
          description: Unique jurisdiction identifier
          readOnly: true
          writeOnly: false
        employeeId:
          type: string
          format: uuid
          description: Id of the employee to whom this jurisdiction belongs
        boundaryRelation:
          type: array
          items: 
            type: string
          description: Boundary Relationship id refered from boundary service
        isActive:
          type: boolean
          default: true
          description: Indicates whether jurisdiction is active
        auditDetail:
          $ref: '#/components/schemas/AuditDetails'

      x-businessRules:
        - Boundary relation must exist in Boundary service